Moved into package io.anserini
