+++ src/main/java/io/anserini/Args.java
@@ -1,4 +1,4 @@
-package luceneingester;
+package io.anserini;
 
 /**
  * Licensed to the Apache Software Foundation (ASF) under one or more
